DSCI creation got stuck in operator reconciliation in selfmanaged clusters (#1813)
With commit ecdb2dc4, we introduced helper functions to retrieve
singleton objects like DataScienceCluster and DSCInitialization.
However, under certain yet-to-be-investigated circumstances, retrieving
DSCInitialization does not always work reliably. This commit aims to
reintroduce the ad-hoc logic for retrieving both DSC and DSCI to
address this issue.

Once the root cause will be determined, we should revert back this code
and use the shared singleton retreieval logic.