Remove pkg/controller/client.Client and use dedicated clients/function
The pkg/controller/client.Client type was introduced during the
refactoring of the operator to separate the DSC reconciliation
logic into dedicated controllers, as it was initially needed to
manage different types of clients. This commit amis to remove
the custom client in favour of the default client provided by
the controller-runtime as the need of custom clients has been
reduced since the initial refactoring.

Refactor client usage in controllers
- Removed the pkg/controller/client.Client and replaced it with
  dedicated clients. this way we can inject custom client when
  neeced (non-cached, backed by a specific cache, etc)
- Refactored various controllers and tests to eliminate the use
  of the custom client wrapper.
- Adjusted resource handling functions to use the standard client.