feat: Relocate kueue manifests to odh-operator
https://issues.redhat.com/browse/RHOAIENG-24289

This applies the configs we expect to be necessary on top of the
external kueue manifests, maintained in config/kueue-configs.

This also changes the way that the batch-user label is applied, since it
will not be managed by the odh-operator when the user installs the kueue
operator outside of odh-operator.

An additional watch on ClusterRole monitors the clusterqueue-viewer-role
since it is externally managed, triggering a reconcile of the
default-kueue if the ClusterRole changes, so that we can keep the
necessary labels appropriately set.

Adds some tests for the new batch-user label management logic.