feat: Enable rhoai build from main branch
This required sigificant changes to the Makefile and a few different
strategies:

- conditionally build different versions of some structs, where there is
  an irreconcilable difference between `main` and `rhoai` branches
  (using build tags)
- maintain a separate overlay of manifests and separate bundle, tracking
  `rhoai` specific changes where necessary.

New directories:

- `bundle.rhoai`: contains the RHOAI bundle (`bundle` continues to
  contain the ODH bundle)
- `config/crd.rhoai`: contains RHOAI specific versions of the generated
  CRDs.
- `config/default.rhoai`: contains RHOAI manifests used to run the
  operator using `make deploy`
- `config/manager.rhoai`: contains RHOAI manager resource definitions
  for `make deploy` and the bundle.
- `config/manifests.rhoai`: contains ClusterServiceVersion definitions
  for RHOAI.
- `config/rbac.rhoai`: contains RHOAI-specific RBAC definitions.
- `samples.rhoai`: contains RHOAI-specific DSC/DSCI examples.
- `webhook.rhoai`: contains RHOAI-specific webhook configuration.

It was necessary to upgrade controller-tools to 0.17.3 to get a
version of `controller-gen` that can use build tags.

With these changes most Make targets now accept the `ODH_PLATFORM_TYPE`
parameter, and operate in either an odh-mode by default, or a rhoai mode
if overridden to any value other than `OpenDataHub`.

`get_all_manifests.sh` now has a different mode when passed
`ODH_PLATFORM_TYPE` other than `OpenDataHub`, where it looks at $VERSION
and infers the downstream git reference to use. (It is most easily
invoked via `make get-manifests ODH_PLATFORM_TYPE=rhoai`).

This adds RHOAI-specific Dockerfiles for the operator and the bundle.

See the difference between the rhoai versions and odh versions by using
a diff tool, such as `meld` or `diff -u`.

You can compare the resulting bundle for differences by checking out
the rhoai branch, and comparing `bundle.rhoai` to `bundle` in the
`rhoai` branch.

There are a number of small differences related to changes that haven't
been made to the `rhoai` branch.