Use custom manifests without devFlags (#2427)
* Use custom manifests without devFlags

Jira: https://issues.redhat.com/browse/RHOAIENG-31642

This gives instructions for an alternative approach to using custom
manifests with the operator, which don't require the use of "devFlags"
in the DSC.

The intention is that this will allow us to remove the devFlags
functionality for 3.0.

This is a primitive approach for now. I expect that folks will start
to build better development patterns rather than needing to modify the
patch file for their component, etc. They can even have their own
approaches in their own repositories, and this may just be an example.

* Don't re-chown files in Dockerfile

They're already chowned as expected from the COPY instruction on L60.