feat(maas): integrate MaaS controller to DSC (#3014)
* feat(maas): integrate ModelsAsService with DataScienceCluster API

Implement DataScienceCluster integration for MaaS component as part of the KServe component.
For now, MaaS requires explicit enablement `managementState: "Managed"` (for technical preview).

Signed-off-by: Edgar Hernández <ehernand@redhat.com>

* feat(maas): integrate ModelsAsService with ODH overlay

Updates ModelsAsService component to use the ODH overlay instead of the OpenShift overlay. The handling of the params.env file is updated with the right parameter names: maas-api-image, gateway-namespace, and gateway-name.

Signed-off-by: Edgar Hernández <ehernand@redhat.com>

* Fix feedback: CI

Signed-off-by: Edgar Hernández <ehernand@redhat.com>

* Fix feedback: Steven

Signed-off-by: Edgar Hernández <ehernand@redhat.com>

* Fix feedback: CI - bring back .gitkeep in opt/manifests

Signed-off-by: Edgar Hernández <ehernand@redhat.com>

---------

Signed-off-by: Edgar Hernández <ehernand@redhat.com>